Move android code to a C file, reduce duplication
